chore: add facade api
